fix: remove lighthouse provider
